Move treemap and Sunburst away temporary
