+++ client/app/charts/index.ts
@@ -2,10 +2,7 @@ export * from './chart.class';
 export * from './bar-chart/bar-chart.component';
 export * from './force-directed-graph/force-directed-graph.component';
 export * from './line-chart/line-chart.component';
-export * from './treemap-chart/treemap-chart.component';
-export * from './sunburst-chart/sunburst-chart.component';
 export * from './hierarchical-edge-bundling/hierarchical-edge-bundling.component';
 export * from './ngx-charts';
 export * from './pie-chart';
-export * from './sunburst-chart';
 export * from './dendogram/dendogram.component';
