+++ client/app/slides/services/charts.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Type, ComponentFactoryResolver } from '@angular/core';
-import { PieChartComponent, AdvancedPieChartComponent, BarChartComponent, ForceDirectedGraphComponent, GaugeChartComponent, 
-  HierarchicalEdgeBundlingComponent, LineChartComponent, SunburstChartComponent, TreemapChartComponent, DendogramComponent,
+import { PieChartComponent, AdvancedPieChartComponent, BarChartComponent, ForceDirectedGraphComponent, GaugeChartComponent,
+  HierarchicalEdgeBundlingComponent, LineChartComponent, DendogramComponent,
   NgGraphComponent
 }  from "app/charts";
 
@@ -14,7 +14,6 @@ export class ChartsService {
     this.listWidget = new Map<string, Type<any>>([['PieChartComponent', PieChartComponent], ['AdvancedPieChartComponent', AdvancedPieChartComponent],
      ['BarChartComponent', BarChartComponent], ['ForceDirectedGraphComponent', ForceDirectedGraphComponent],
      ['GaugeChartComponent', GaugeChartComponent], ['HierarchicalEdgeBundlingComponent', HierarchicalEdgeBundlingComponent],
-     ['SunburstChartComponent', SunburstChartComponent], ['TreemapChartComponent', TreemapChartComponent],
      ['LineChartComponent', LineChartComponent], ['DendogramChartComponent', DendogramComponent], ['NgGraphComponent', NgGraphComponent]
      ]);
   }
